purism/librem_14: Bring back 1.7 PL4 logic
Restore 1.7 implementation of power_peci_limit(), which sets PL4 every
iteration with up to 16 retries.

Testing has shown that this is a source of sudden shutoffs and that
reverting to the 1.7 version mitigates them.  While we are still
investigating the root cause, bring back the 1.7 implementation to
address issues in the field.

Signed-off-by: Jonathon Hall <jonathon.hall@puri.sm>